pyproject.toml
[project]
name = "securenetops-dockerlab"
version = "0.1.0"
description = "Secure NetOps examples (Docker network lab)"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "PyYAML>=6.0.2",
  "pydantic>=2.8.0",
  "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
